+++ .eslintrc
@@ -28,7 +28,7 @@
   "overrides": [
     {
       "files": [
-        "tests/unit/*.spec.js"
+        "tests/unit/**/*.spec.js"
       ],
       "env": {
         "jest": true
